Implemented value checking for new transaction

+++ app/src/main/java/com/example/abdullah/budgetary/piechart/PieSliceInterface.java
@@ -1,7 +1,7 @@
 package com.example.abdullah.budgetary.piechart;
 
 public interface PieSliceInterface {
-    double getValue(int index);
+    Long getValue(int index);
     long getId(int index);
     String getLabel(int index);
     int getColor(int index);
